Initial worker queue & more modular code.
