+++ rust_bot/src/message_utils.rs
@@ -1,4 +1,3 @@
-use crate::read_and_write::Summary;
 use serde::Serialize;
 use std::collections::HashMap;
 use time::{Date, OffsetDateTime, Time};
